% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cookies.R
\docType{data}
\name{js_cookies}
\alias{js_cookies}
\title{https://gist.github.com/calligross/e779281b500eb93ee9e42e4d72448189
js_cookies}
\format{
An object of class \code{character} of length 1.
}
\usage{
js_cookies
}
\description{
https://gist.github.com/calligross/e779281b500eb93ee9e42e4d72448189
js_cookies
}
\keyword{datasets}
