% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{multiple_checkbox}
\alias{multiple_checkbox}
\title{multiple_checkbox}
\usage{
multiple_checkbox(
  input_id,
  label,
  choices,
  selected = NULL,
  position = "grouped",
  type = "radio",
  ...
)
}
\description{
multiple_checkbox
}
