% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_mod.R
\name{theme_ui}
\alias{theme_ui}
\title{theme_ui
https://stackoverflow.com/questions/19844545/replacing-css-file-on-the-fly-and-apply-the-new-style-to-the-page}
\usage{
theme_ui(id)
}
\description{
theme_ui
https://stackoverflow.com/questions/19844545/replacing-css-file-on-the-fly-and-apply-the-new-style-to-the-page
}
